Refactor user components
